Fix for not responsive home screen widget
If the app is killed the widget doesn't receive updates
and it keeps the old value of the editor.
Pressing f.e. cursor buttons in such case
has not effect as real editor is empty and no
"editor change" events are triggered, thus,
widget is not updated.

To solve the issue let's just update the widget
any time its buttons are pressed.